+++ lib/testScenario/postmanCollectionGenerator.ts
@@ -30,6 +30,10 @@ export interface PostmanCollectionGeneratorOption
   generateCollection: boolean;
   baseUrl: string;
   validationLevel?: ValidationLevel;
+  cleanUp: boolean;
+  from?:string,
+  to?:string,
+  runId?:string
 }
 
 @injectable()
@@ -55,7 +59,7 @@ export class PostmanCollectionGenerator {
       }
     }
     let index = 0;
-    const runId = generateRunId();
+    const runId = this.opt.runId || generateRunId();
     if (this.opt.markdownReportPath) {
       await this.fileLoader.writeFile(this.opt.markdownReportPath, generateMarkdownReportHeader());
     }
@@ -75,17 +79,21 @@ export class PostmanCollectionGenerator {
         jsonLoader: this.testResourceLoader.jsonLoader,
         baseUrl: this.opt.baseUrl,
         validationLevel: this.opt.validationLevel,
+        from : this.opt.from,
+        to: this.opt.to
       };
 
       const client = inversifyGetInstance(PostmanCollectionRunnerClient, opts);
-      const runner = new TestScenarioRunner({
-        jsonLoader: this.testResourceLoader.jsonLoader,
-        env: this.env,
-        client: client,
-      });
+       const runner = new TestScenarioRunner({
+         jsonLoader: this.testResourceLoader.jsonLoader,
+         env: this.env,
+         client: client,
+       });
       await runner.executeScenario(testScenario);
       // If shared resource-group, move clean to one separate scenario.
-      await runner.cleanAllTestScope();
+      if (this.opt.cleanUp) {
+        await runner.cleanAllTestScope();
+      }
       for (let i = 0; i < client.collection.items.count(); i++) {
         this.longRunningOperationOrderUpdate(client, i);
       }